chore: CR refactors
